Move prose.* finding to html-require-recipe-ingredients
